[MDH-105] refactor : 메모리 누수 문제로 인해 DTO 객체 외부 클래스로 변경
